Made ready for npm publish
