+++ package.json
@@ -1,8 +1,9 @@
 {
   "name": "extract-cbd-shape",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "description": "Extract an entity based on CBD and a SHACL shape",
-  "main": "lib/extract-cbd-shape.js",
+  "main": "dist/lib/extract-cbd-shape.js",
+  "types": "dist/lib/extract-cbd-shape.d.ts",
   "scripts": {
     "test": "mocha tests/*",
     "build": "tsc"
