+++ tests/01 - fetching a shacl shape/extraction.test.ts
@@ -1,6 +1,6 @@
 import { assert } from "chai";
 import {Store, NamedNode, Writer, Term, Parser, StreamParser} from "n3";
-import {CBDShapeExtractor} from '../../lib/extract-cbd-shape';
+import {CBDShapeExtractor} from '../../lib/CBDShapeExtractor';
 import rdfDereference from 'rdf-dereference';
 
 describe('Check whether we can successfully extract a SHACL shape', async () => {
