+++ tests/04 - logical edge cases/testShapeTemplate.test.ts
@@ -1,6 +1,6 @@
 import { assert } from "chai";
 import {Store, NamedNode, Writer, Term, Parser, StreamParser} from "n3";
-import {Shape, ShapesGraph} from '../../lib/Shape';
+import {ShapeTemplate, ShapesGraph} from '../../lib/Shape';
 import rdfDereference from 'rdf-dereference';
 
 describe('Test shape template of the logical edge cases', function () {
