+++ tsconfig.json
@@ -1,1 +1,12 @@
-{}
+{
+    "compilerOptions": {
+      "declaration": true,
+      "outDir": "dist",
+      "esModuleInterop": true
+    },
+    "exclude": [
+      "node_modules",
+      "dist",
+      "tests"
+    ]
+  }
